Fix: icon trying to load from file when empty icon name
File.new_for_path("") will just resolve to the working directory, and as that exists, it tries to load the working directory as icon.

I also changed path detection from `.contains ("/")` to `.has_prefix ("/")` as I think we always have absolute paths. And I'm not sure if theme icon names could contain a '/'.