Improve field validation behaviour & icon validation
In some edge cases, the UI doesn't show if fields are invalid - so the user doesn't know why the Save button is greyed out.

That's the UX problem of e.g. #85

I refactored so that Form field error status is updated in .validate (now called 'updateFormStatus'), and the change listeners only set the validity boolean.

Also update the icon back to default if the name field is cleared.

I also added validation feedback for the image button (although it feels weird to set class 'destructive-action', I found it to be the easiest :P)